+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Implement support for F# projects (.fsproj)
 
 ## [0.2.1] / 2023-11-22
 - Normalize package ids using ToLowerInvariant before calling NuGet api